ICU_DECODE causes the URL parts to be unescaped. For our case, this can be catastrophic if the url contains a + character and the server is hosted on an IIS web server. IIS (and certain azure function hosts) do not route well resources that have the "+" character unless duble-escaping is enabling on the server which is not a default setting. When this happens, the server might return a 404 or a 403 depending on the unescaped resource and the routing configuration of the server.
This change removed the ICU_DECODE flag from the simple_request implementation.
We verified with our WinHTTP experts there is no good reason to be doing ICU_DECODE before sending out a request through winHTTP or winINET.
The only case in which it might make sense if the server does not expect escaped URLs
(which would be a failure to implement the standard) or if the Url is being craked for parsing but not for sending.