Libzip 1.7.3 compatibility fix for the Omaha Open Source build.
Version 1.7.3 needs an additional generated file zip_err_str.c. This
file is missing in the distribution itself and has to be generated.
To make it easier for people to build open source without having to
invoke CMake and generate this file, this file is being included
in this CL.