Add Enterprise detection of MDM enrolled machines.
This change allows MDM enrolled machines that are running a non-personal version of Windows to be considered domain-joined machines. Note: MDM is not the same as Azure Active Directory, which will be a separate change.

Also fixed an issue where a couple of policies were not falling back to DM policies.