Improve the user experience when downloading large files in Omaha.
This change contains two work arounds to slightly improve the user experience due to long times taken by hash verification of large files (>512MB).

First, it changes the text displayed when the download is complete from "Download complete." to "Download complete. Please wait while the installer is verified.". The change is made for en-us only. GRIT in Omaha is broken at the moment. i18n must be dealt with in a follow up CL.

Second, checking of hashes occurs at several times during /install execution.
In particular, it happens once right before running the installer. Unfortunately, during the hash check, the model lock in Omaha is taken, so Omaha can't respond to polling for state changes. That means that the UI keeps displaying "Download Complete." instead of displaying "Waiting to install...". There is no easy workaround for this. A delay is introduced in the WaitingToInstall state to allow the UI client to catch up with the state transition, and show the "Waiting to install..." text while the hash is checked.