Fix broken WebServicesClientTest.SendUsingCup test.
The prod is not responding with Etag any longer. We are using
the kHeaderXETag exclusively now.

Handling this header is encapsulated in the CupEcdsaRequest
and not visible at all in the WebServiceClient class.

The test determines that sending with CUP succeeds solely
by instantiating the WebServiceClient with `use_cup=true`.

There is no need to check for the presence of CUP headers,
which were not tested in the past either.