Apply wait before DM registration.
This change adds a jitter that can be up to 1 minute, which will help the issue, which, together with random boots of client machines may alleviate the issue. 1 minute is a compromise because we do not want to wait too long in UpdateApps.

Policy fetch does not currently need the jitter. Registration is more heavy-weight at the moment.