Display Progress Text in the Title Bar.
Before this change, the title bar text was static. This change allows for the progress to be seen in the top level title bar, which helps screen readers among other things.

Because of issues with grit and python, I decided to reuse an existing string translation, IDS_APPLICATION_NAME_CONCATENATION. Not ideal, in the sense that it uses a comma for separation instead of a dash, but good enough.