Use CreateProcess instead of ShellExecute for background process launches.
Crashing while calling ShellExecuteEx is the 2nd largest bucket in Omaha, hence the change to avoid using ShellExecute unless the process being launched is a foreground process. ShellExecute is needed for foreground processes, since it does a good job of moving the UX keyboard focus to the new process, but not for background processes.