Fix Downgrade to a more stable channel via TargetChannel does not work.
Now rollback_allowed is sent in the updatecheck independent of whether the TargetVersionPrefix policy is set.

With this change, with RollbackToTargetVersion enabled, even if no "Target version prefix override" is set, but the version available is lower than the installed version on the client, a rollback will happen in this case as well. This can happen for instance if the "Target Channel override" is changed from Dev or Beta (higher versions) to Beta or Stable (lower versions).

If RollbackToTargetVersion is not configured or is disabled, installs that have a version higher than that specified by "Target version prefix override", or in the current Channel, will be left as-is.

If RollbackToTargetVersion is enabled, installs that have a version higher than that specified by "Target version prefix override", or in the current Channel,  will be downgraded to the highest available version that matches the target version, or the current Channel, respectively.