Fix Security Arbitrary file read when caching file CallAsSelfAndImpersonate2
We now open the downloaded file as the current (impersonated) user. This ensures that we are not reading any privileged files that are otherwise inaccessible to the impersonated user.  We then hand-copy the file to the Package Cache unimpersonated, since the package cache is in a privileged location.# Please enter the commit message for your changes. Lines starting