BITS: More logging and fallback faster on error
There are enterprise customers who report issues with BITS jobs that are stuck with transient errors, or have questions about jobs in the BITS queue.

This CL sets the BITS "no progress" timeout to 0 for all jobs to prevent retries and forces the jobs into the BG_JOB_STATE_ERROR state immediately if any errors are encountered.

This CL also adds OPT logging in the Omaha log from the BITS request each time we get the state of the request, which will help with customers who report BITS issues.