Revert sending the current version "pv" for over-installs
This CL is causing some rule breakage in the update configs, because we are sending the current version "pv" in update checks with a '-' in front of the `current_version`. Reverting this CL for now, we will revisit with a better change if there is enough demand for the feature.