Fix Take Group Policy Critical Section to prevent race conditions
In the past, an attempt was made to fix this bug by taking the Group
Policy critical section for all Enterprise-joined cases at startup.
This caused deadlocks when the installer was run via GPO. crbug/1231513.
When the install is pushed via GPO, GPO takes the critical section, and
we hang at the ::EnterCriticalPolicySection(TRUE) call.

We now initialize the policy managers without taking the Group Policy
critical section. Later, once we know the exact scenario that we are
operating under, we may reload the policies with the critical section
lock. At the moment, we reload the policies with the critical section
lock for all User installs and updates, as well as all Machine updates.

For Machine installs, we do not reload policies for fresh installs,
because there is the possibility that we are running under GPO, and
because GPO takes the critical section itself, this can result in a
deadlock. Also, Machine installs require Admin, and an Admin can bypass
Group Policy if they wish to, so we are ok with sacrificing some
accuracy in terms of policy correctness in return for not having
deadlocks.

Updates do not run under GPO, so we reload the policies with the Group
Policy critical section for both User and Machine updates.