Build protoc for Omaha using Scons builders.
This CL adds logic to the Omaha base build definition to build protoc,
replacing the use of gyp for the Omaha build.

In order to simplify management of the list of source files, I added
a helper function that is similar to Blaze's go/be#glob function.
The Omaha build does not distinguish the lite protobuf runtime from
the full runtime, so the correct files are fairly straightforward to
find by pattern matching.

There are some slight differences from the previous setup:

- protoc.exe is built using a Scons builder. It uses the libprotobuf
  library that is already built and used by crx_file.
- Warnings are lowered from `/W4` to `/W2`, which reduces the number
  of `/wdNNNN` flags.
- libprotobuf:
    - Well-known types (any.pb.cc, time.pb.cc, etc.) are now linked in
    - A few more source files are also linked in, but should remain
      unused
- The protoc importer moved from libprotobuf to protoc. The importer
  allows runtime parsing of `.proto` files, and can be built without
  the rest of the protocol compiler (such as the command-line
  interface and code generators). However, it does not appear to be
  needed by crx_file, so it is safe to move to protoc.

Overall, I expect this will make the Omaha build more resilient to
future changes in the protobuf library.