Add support for custom application logo in the Omaha progress UI
This CL uses `::OleLoadPicturePath` to load the application logo from a URL and shows it in the Omaha progress UI. The url for the logo is in BMP format, and is deduced based on the `app_id_`.

The app logos are hosted at `{kUrlAppLogo}{url escaped app_id}.bmp`. If `{url escaped app_id}.bmp` exists, a logo is shown in the updater UI for that app install.

`kUrlAppLogo` is specified in omaha/base/const_addresses.h.