Use GoogleUpdate.exe itself when elevating machine handoff installs
This CL stops copying the `metainstaller` to the versioned install directory, and instead uses `GoogleUpdate.exe` within the same secure `%programfiles%` directory for elevating a machine install.

The `metainstaller` was previously needed for cross-install scenarios, i.e., a user installation `/handoff` doing a machine install, where elevating the entire `metainstaller` was safer. However, that scenario was only used by the ActiveX/plugin code, and is no longer in use.