Revert to using the metainstaller for elevating machine handoff installs
The previous CL used `GoogleUpdate.exe` for elevating machine installs. However, this made machine installs elevate from the `TMP` directory, which is insecure. This CL reverts to using the metainstaller for elevating machine handoff installs, as it was before.

This CL does keep the setup changes of not copying the metainstaller to the versioned install directory. The metainstaller was previously needed for cross-install scenarios, i.e., a user installation /handoff doing a machine install, where elevating the entire metainstaller was safer. However, that scenario was only used by the ActiveX/plugin code, and is no longer in use.