Fix Google Update leaves empty folders after uninstalling
This CL changes GoogleUpdate to use `%systemroot%\SystemTemp` for the
crash reporting and as the temp folder for certain setup operations,
instead of creating directories under %programfiles%\google, if the
former directory is available on the system.

`SystemTemp` is available on most recent versions of Windows 10 and 11.
`SystemTemp` is secure and only accessible to admins and local system.

In order to have continued support for older versions of Windows, this
CL retains a fallback to the original behavior of using %programfiles%.