More fixes for GoogleUpdate leaves empty folders after uninstalling
This CL changes the metainstaller to also use `%systemroot%\SystemTemp` for the secure temp folder for setup operations instead of creating directories under `%programfiles%\google`, if the former directory is available on the system.

`SystemTemp` is available on most recent versions of Windows 10 and 11. `SystemTemp` is secure and only accessible to admins and local system.

In order to have continued support for older versions of Windows, this CL retains a fallback to the original behavior of using `%programfiles%`.

This CL has inlined several functions in `utils.h` and `file.h`, so that the metainstaller can call these functions, since the metainstaller does not link to `base.lib`.