Fix Updates not suppressed when the duration rolls over to the next day
This CL fixes this issue where if the `UpdatesSuppressedStartHour` + `UpdatesSuppressedStartMin` + `UpdatesSuppressedDurationMin` rolls over to the next day, then the updates would start being applied at the first update check after midnight, even if the duration is not up yet.

So for instance, if the policy is to suppress updates `from 6 PM for 12 hours`, the updates would start getting downloaded and installed starting at `midnight`, even though the updater is supposed to `wait until 6 AM` the next day.