Revert to using `::SHQueryValueEx` instead of `::RegQueryValueEx`
https://www.geoffchappell.com/studies/windows/shell/shlwapi/api/reg/queryvalueex.htm
```
Although this function and the standard API function RegQueryValueEx have exactly the same prototype, the two do not behave identically. A call to the SHLWAPI function is essentially a call to the standard function but with post-processing, presumably with the idea of improving or even correcting the standard function. There are two general aims:

to fix REG_SZ or REG_EXPAND_SZ data that is not properly null-terminated;
to expand REG_EXPAND_SZ data and return it as REG_SZ data.
```