Fix Elevation of Privilege in GoogleUpdate related to impersonation
When GoogleUpdate runs Code Red, and a Code Red file is served by the
server (or a malicious MitM), the GoogleUpdate code was under some
circumstances doing a Move/Delete of a file that could be under a
Medium integrity user's control with Local System privileges.
This CL fixes this issue.